requirements.txt
# Core
python-dotenv==1.0.1

# Neo4j & Graph DB
neomodel==5.1.0

# API
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.1  # Added explicit dependency for Pydantic models

# LLM Integration
requests==2.32.3

# Dev & Debug
httpx==0.27.0  # For FastAPI testing
black==24.4.2
isort==5.13.2

# Optional: ANTLR runtime if using from Python
antlr4-python3-runtime==4.13.1

# Testing
pytest==8.4.0